requirements.txt
argcomplete==1.12.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
commitizen==2.21.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decli==0.5.2
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
questionary==1.10.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomlkit==0.10.0
typing-extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
